Add random seeding.
Provide a random seed to each act call (likely unique). When debugging,
simply ensure that the random module is seeded with the game provided
seed in game.seed for each bot in order to have reproducible results.

Obviously not easily enforcible on the server, but this is more for
local debugging where an exception or behaviour is only exhibited in a
limited number of games.